+++ deep_quoridor/src/renderers/match_results.py
@@ -1,0 +1,11 @@
+from renderers import Renderer
+from arena import GameResult
+from agents import Agent
+
+
+class MatchResultsRenderer(Renderer):
+    def start_game(self, game, agent1: Agent, agent2: Agent):
+        self.match = f"{agent1.name()} vs {agent2.name()}"
+
+    def end_game(self, game, result: GameResult):
+        print(f"{result.game_id}: {self.match} - {result.winner} won in {result.steps} steps")
