+++ deep_quoridor/src/renderers/none.py
@@ -1,0 +1,5 @@
+from renderers import Renderer
+
+
+class NoneRenderer(Renderer):
+    pass
